move ffmpeg stuff to util/
