+++ Makefile
@@ -15,4 +15,4 @@ make-screenshot:
 	go build ./test/make-screenshot
 
 count:
-	wc -l `find cmd -name "*.go"` `find lib -name "*.go"` `find test -name "*.go"`
+	wc -l `find cmd -name "*.go"` `find lib -name "*.go"` `find test -name "*.go"` `find util -name "*.go"`
