+++ cmd/nsf/main.go
@@ -4,19 +4,17 @@ import (
     "log"
     "fmt"
     "os"
-    "io"
     "context"
     "time"
     "strings"
     "sync"
     "path/filepath"
     "strconv"
-    "os/exec"
-    "syscall"
     "bytes"
     "encoding/binary"
 
     nes "github.com/kazzmir/nes/lib"
+    "github.com/kazzmir/nes/util"
     "github.com/veandco/go-sdl2/sdl"
 
     "github.com/jroimartin/gocui"
@@ -406,50 +404,6 @@ func convertTime(data string) (uint64, error) {
     return number * uint64(multiple), nil
 }
 
-/* FIXME: share this between the nes program */
-func findFfmpegBinary() (string, error) {
-    return exec.LookPath("ffmpeg")
-}
-
-func waitForProcess(process *os.Process, timeout int){
-    done := time.Now().Add(time.Second * time.Duration(timeout))
-    dead := false
-    for time.Now().Before(done) {
-        err := os.Signal(syscall.Signal(0)) // on linux sending signal 0 will have no impact, but will fail
-                            // if the process doesn't exist (or we don't own it)
-        if err == nil {
-            time.Sleep(time.Millisecond * 100)
-        } else {
-            dead = true
-            break
-        }
-    }
-    if !dead {
-        /* Didn't die on its own, so we forcifully kill it */
-        log.Printf("Killing pid %v", process.Pid)
-        process.Kill()
-    }
-    process.Wait()
-}
-
-func niceSize(path string) string {
-    info, err := os.Stat(path)
-    if err != nil {
-        return ""
-    }
-
-    size := float64(info.Size())
-    suffixes := []string{"b", "kb", "mb", "gb"}
-    suffix := 0
-
-    for size > 1024 && suffix < len(suffixes) - 1 {
-        size /= 1024
-        suffix += 1
-    }
-
-    return fmt.Sprintf("%.2f%v", size, suffixes[suffix])
-}
-
 func saveMp3(nsfPath string, mp3out string, track int, renderTime uint64) error {
     nsf, err := nes.LoadNSF(nsfPath)
     if err != nil {
@@ -477,114 +431,8 @@ func saveMp3(nsfPath string, mp3out string, track int, renderTime uint64) error 
 
     waiter.Add(1)
     go func(){
-        ffmpeg_binary_path, err := findFfmpegBinary()
-        if err != nil {
-            log.Printf("Could not find ffmpeg: %v", err)
-            return
-        }
-
-        audio_reader, audio_writer, err := os.Pipe()
-        if err != nil {
-            return
-        }
-
-        log.Printf("Launching ffmpeg")
-        ffmpeg_process := exec.Command(ffmpeg_binary_path,
-            "-use_wallclock_as_timestamps", "1", // treat the incoming data as a live stream
-            "-f", "f32le", // audio is uncompressed pcm in float32 format
-            "-ar", strconv.Itoa(int(sampleRate)), // sample rate
-            "-ac", "1",
-            "-i", "pipe:3", // audio is passed as fd 3
-
-            "-tune", "zerolatency", // fast encoding
-            "-acodec", "mp3", // mp3 for audio
-            // "-filter:a", "volume=10dB", // increase volume a bit
-            "-y", // overwrite output if the file already exists
-            mp3out)
-
-        // ffmpeg_process.Stdin = reader
-
-        /* FIXME: figure out a solution for windows */
-        ffmpeg_process.ExtraFiles = []*os.File{audio_reader}
-
-        stdout, err := ffmpeg_process.StdoutPipe()
-        if err != nil {
-            log.Printf("Could not get ffmpeg stdout: %v", err)
-            return
-        }
-
-        stderr, err := ffmpeg_process.StderrPipe()
-        if err != nil {
-            log.Printf("Could not get ffmpeg stderr: %v", err)
-            return
-        }
-
-        err = ffmpeg_process.Start()
-        if err != nil {
-            log.Printf("Could not start ffmpeg: %v", err)
-            return
-        }
-
-        go func(stdout io.ReadCloser){
-            buffer := make([]byte, 4096)
-            for {
-                count, err := stdout.Read(buffer)
-                if err != nil {
-                    log.Printf("Could not read ffmpeg stdout: %v", err)
-                    return
-                }
-                _ = count
-                // log.Printf("ffmpeg: %v", string(buffer[0:count]))
-            }
-        }(stdout)
-
-        go func(stdout io.ReadCloser){
-            buffer := make([]byte, 4096)
-            for {
-                count, err := stdout.Read(buffer)
-                if err != nil {
-                    log.Printf("Could not read ffmpeg stdout: %v", err)
-                    return
-                }
-                _ = count
-                // log.Printf("ffmpeg: %v", string(buffer[0:count]))
-            }
-        }(stderr)
-
-        log.Printf("Recording to %v", mp3out)
-
-        go func(){
-            defer waiter.Done()
-            startTime := time.Now()
-            <-quit.Done()
-            /* ffmpeg will normally close on its own if its input is closed */
-            audio_writer.Close()
-            /* we can send SIGINT to it as well, which also usually stops it */
-            ffmpeg_process.Process.Signal(os.Interrupt)
-            waitForProcess(ffmpeg_process.Process, 10)
-            log.Printf("Recording has ended. Saved '%v' for %v size %v", mp3out, time.Now().Sub(startTime), niceSize(mp3out))
-        }()
-
-        go func(){
-            defer audio_reader.Close()
-
-            var audioBuffer bytes.Buffer
-
-            for {
-                select {
-                    case <-quit.Done():
-                        return
-                    case audio := <-audioOut:
-                        audioBuffer.Reset()
-                        /* convert []float32 into []byte */
-                        for _, sample := range audio {
-                            binary.Write(&audioBuffer, binary.LittleEndian, sample)
-                        }
-                        // log.Printf("Enqueue audio")
-                        audio_writer.Write(audioBuffer.Bytes())
-                }
-            }
-        }()
+        defer waiter.Done()
+        util.EncodeMp3(mp3out, quit, int(sampleRate), audioOut)
     }()
 
     log.Printf("Rendering track %v of %v to '%v' for %d:%02d", track+1, filepath.Base(nsfPath), mp3out, renderTime/60, renderTime % 60)
